Adds test_inst, simply for example purposes.
This change is meant to serve as an example of how to add a very
specific type of instruction. test_inst is a two-byte opcode which
is decoded based on an immediate value, which is how other gem5
pseudoops are decoded. By grouping them with the other pseudoops,
and modifying the appropriate m5op* files, we can generate our op
easily from our own C++ code, which is useful for testing.

Note that I'm not actually using this approach going forward, as
I would prefer to reserve a one-byte opcode and not use any
immediate value for decoding. I'd prefer to keep the format open,
in case I get to the point where I can add arguments to my custom
instructions.

Change-Id: Ic5ed6095825a950fa4eb91f9cb66fb6bd55fab1d